requirements.txt
# SDLC SimLab - Python Dependencies

# Core Simulation
# mesa>=2.1.0  # Agent-based modeling framework (uncomment when ready)
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0

# Configuration
pyyaml>=6.0.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
websockets>=12.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0  # Async PostgreSQL driver
alembic>=1.12.0

# Time-series (optional, for later)
# influxdb-client>=1.38.0

# Task Queue
celery>=5.3.0
redis>=5.0.0
redis[asyncio]>=5.0.0  # Async Redis for WebSocket pub/sub

# External Integrations
httpx>=0.25.0
PyGithub>=2.1.0
python-gitlab>=4.0.0

# Data Processing
python-dateutil>=2.8.0
pytz>=2023.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
ruff>=0.1.0
black>=23.0.0
mypy>=1.6.0

# Development
ipython>=8.16.0
python-dotenv>=1.0.0
